Moved servers to a dedicated package
